es6 module init
