+++ backbone.mjs
@@ -1,0 +1,8 @@
+// Allow the `Backbone` object to serve as a global event bus, for folks who
+// want global "pubsub" in a convenient place.
+// _.extend(Backbone, Events);
+
+export * from './events.mjs';
+export * from './view.mjs';
+export * from './collection.mjs';
+
