+++ helpers.mjs
@@ -1,0 +1,52 @@
+import _ from 'lodash';
+
+// Helpers
+// -------
+
+// Helper function to correctly set up the prototype chain for subclasses.
+// Similar to `goog.inherits`, but uses a hash of prototype properties and
+// class properties to be extended.
+export var extend = function(protoProps, staticProps) {
+    var parent = this;
+    var child;
+
+    // The constructor function for the new subclass is either defined by you
+    // (the "constructor" property in your `extend` definition), or defaulted
+    // by us to simply call the parent constructor.
+    if (protoProps && _.has(protoProps, 'constructor')) {
+        child = protoProps.constructor;
+    } else {
+        child = function(){ return parent.apply(this, arguments); };
+    }
+
+    // Add static properties to the constructor function, if supplied.
+    _.extend(child, parent, staticProps);
+
+    // Set the prototype chain to inherit from `parent`, without calling
+    // `parent`'s constructor function and add the prototype properties.
+    child.prototype = _.create(parent.prototype, protoProps);
+    child.prototype.constructor = child;
+
+    // Set a convenience property in case the parent's prototype is needed
+    // later.
+    child.__super__ = parent.prototype;
+
+    return child;
+};
+
+// Set up inheritance for the model, collection, router, view and history.
+// Model.extend = Collection.extend = Router.extend = View.extend = History.extend = extend;
+
+// Throw an error when a URL is needed, and none is supplied.
+export var urlError = function() {
+    throw new Error('A "url" property or function must be specified');
+};
+
+// Wrap an optional error callback with a fallback error event.
+export var wrapError = function(model, options) {
+    var error = options.error;
+    options.error = function(resp) {
+        if (error) error.call(options.context, model, resp, options);
+        model.trigger('error', model, resp, options);
+    };
+};
